+++ internal/backend/remote-state/s3/backend_complete_test.go
@@ -206,8 +206,7 @@ aws_secret_access_key = ProfileSharedCredentialsSecretKey
 
 		"environment AWS_ACCESS_KEY_ID does not override config Profile": {
 			config: map[string]any{
-				"profile":             "SharedCredentialsProfile",
-				"use_legacy_workflow": false,
+				"profile": "SharedCredentialsProfile",
 			},
 			EnvironmentVariables: map[string]string{
 				"AWS_ACCESS_KEY_ID":     servicemocks.MockEnvAccessKey,
@@ -235,7 +234,8 @@ aws_secret_access_key = ProfileSharedCredentialsSecretKey
 
 		"environment AWS_ACCESS_KEY_ID overrides config Profile in legacy workflow": { // Legacy behavior
 			config: map[string]any{
-				"profile": "SharedCredentialsProfile",
+				"profile":             "SharedCredentialsProfile",
+				"use_legacy_workflow": true,
 			},
 			EnvironmentVariables: map[string]string{
 				"AWS_ACCESS_KEY_ID":     servicemocks.MockEnvAccessKey,
@@ -254,7 +254,11 @@ aws_secret_access_key = DefaultSharedCredentialsSecretKey
 aws_access_key_id = ProfileSharedCredentialsAccessKey
 aws_secret_access_key = ProfileSharedCredentialsSecretKey
 `,
-			ValidateDiags: ExpectNoDiags,
+			ValidateDiags: ExpectDiagMatching(
+				tfdiags.Warning,
+				equalsMatcher("Deprecated Parameter"),
+				ignoreMatcher{},
+			),
 		},
 
 		"environment AWS_ACCESS_KEY_ID": {